Titles on Library Screen
- Unpacking epub immediately
- Unfortunately still keeping around .epub files to utilize existing
  methods that parse .epub files.
- Will need to write my own method to handle parsing files

+++ SkimReader/SkimReaderViewController.m
@@ -63,7 +63,7 @@ static const NSInteger defaultWPM = 300;
 #pragma mark - Utility
 
 - (void)setWPMValue:(NSInteger)wpm {
-	self.wpmLabel.text = [NSString stringWithFormat:@"%ld WPM", wpm];
+	self.wpmLabel.text = [NSString stringWithFormat:@"%ld WPM", (long)wpm];
 	self.spritzViewController.wordsPerMinute = wpm;
 }
